Add throttling logic
